Add Trips page [WIP]
